Updated centrality variable types

+++ StRoot/StEStructPool/macros/doEStruct2pt.C
@@ -1,5 +1,5 @@
 /************************************************************************
- * $Id: doEStruct2pt.C,v 1.2 2003/11/21 06:26:40 porter Exp $
+ * $Id: doEStruct2pt.C,v 1.3 2004/04/15 18:46:31 msd Exp $
  *
  * Author: Jeff Porter 
  *
@@ -33,7 +33,7 @@ void doEStruct2pt(const char* fileListFile, const char* outputDir, const char* j
   // simple (global) centrality definition ...not persistant to event file.. 
   // and not used in this particular example
   StEStructCentrality* cent=StEStructCentrality::Instance();
-  const int temp[4]={0,4,7,50};
+  const double temp[4]={0,4,7,50};
   cent->setCentralities(temp,4);
 
   // create the low-level reader (here for MuDst)
@@ -129,6 +129,9 @@ void doEStruct2pt(const char* fileListFile, const char* outputDir, const char* j
 /**********************************************************************
  *
  * $Log: doEStruct2pt.C,v $
+ * Revision 1.3  2004/04/15 18:46:31  msd
+ * Updated centrality variable types
+ *
  * Revision 1.2  2003/11/21 06:26:40  porter
  * macros for running pythia
  *
